Refactor CallVisualizer components environments initialization
